requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0

# Azure Speech
azure-cognitiveservices-speech==1.40.0

# OpenAI
openai==1.50.0

# Audio processing
pydub==0.25.1

# Environment
python-dotenv==1.0.1

# Utilities
pydantic==2.9.0
pydantic-settings==2.5.0
